client/requirements.txt
# Core dependencies
Pillow>=9.0.0
Flask>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
omni-epd>=1.5.0
gtfs-realtime-bindings>=1.0.0
protobuf>=4.0.0
# Raspberry Pi hardware
RPi.GPIO>=0.7.0; sys_platform == 'linux'
spidev>=3.6; sys_platform == 'linux'
# omni-epd must be installed from git: pip install git+https://github.com/robweber/omni-epd.git
# Or use Waveshare drivers directly: pip install ~/e-Paper/RaspberryPi_JetsonNano/python/

# Development
pytest>=7.4.0
pytest-mock>=3.12.0
